Delete fortune-demo/staticfortune demo
