+++ .github/workflows/fortune-demo.yml
@@ -32,8 +32,4 @@ jobs:
       - name: Run 'fortune-gradle'
         run: |
           cd fortune-demo/fortune-gradle
-          ./build.sh    
-      - name: Run 'staticfortune'
-        run: |
-          cd fortune-demo/staticfortune
           ./build.sh
